Add files via upload
The Heating, Ventilation & Air Conditioning tile is coded different compared to the other tiles.  The "estimateYN" variable is hidden, so the array is not accessible.   Other tiles use "estimateYN[0]".  This tile must use "estimateYN" so that the work order description is output properly in the "displayDescription", "extraDescription", and "mappingWorkOrderDesc" fields.  

Another issue is the Workflow had to be updated so that the "Bypass (Notifyc FC)" used the correct Method Notification "Notify Only", instead of "Notify with Step"